move test to feature base
