+++ featuretools/tests/primitive_tests/test_agg_feats.py
@@ -35,7 +35,6 @@ from featuretools.primitives import (
 from featuretools.primitives.base import AggregationPrimitive
 from featuretools.synthesis.deep_feature_synthesis import (
     DeepFeatureSynthesis,
-    check_stacking,
     match,
 )
 from featuretools.tests.testing_utils import backward_path, feature_with_name, to_pandas
@@ -47,21 +46,6 @@ def reset_dfs_cache():
     feature_cache.enabled = False
     feature_cache.clear_all()
 
-
-@pytest.fixture
-def test_agg_primitive():
-    class TestAgg(AggregationPrimitive):
-        name = "test"
-        input_types = [ColumnSchema(semantic_tags={"numeric"})]
-        return_type = ColumnSchema(semantic_tags={"numeric"})
-        stack_on = []
-
-        def get_function(self, agg_type="pandas"):
-            return None
-
-    return TestAgg
-
-TODO: add transform fixture
 
 
 def test_get_depth(es):
@@ -199,84 +183,6 @@ def test_mean_nan(es):
         primitive=Mean(skipna=False),
     )
     assert include_nan_feat.get_name() == "MEAN(log.value, skipna=False)"
-
-
-def test_base_of_and_stack_on_heuristic(es, test_aggregation_primitive):
-    child = Feature(
-        es["sessions"].ww["id"],
-        parent_dataframe_name="customers",
-        primitive=Count,
-    )
-    test_aggregation_primitive.stack_on = []
-    child.primitive.base_of = []
-    assert not check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = []
-    child.primitive.base_of = None
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = []
-    child.primitive.base_of = [test_aggregation_primitive]
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = None
-    child.primitive.base_of = []
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = None
-    child.primitive.base_of = None
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = None
-    child.primitive.base_of = [test_aggregation_primitive]
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = [type(child.primitive)]
-    child.primitive.base_of = []
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = [type(child.primitive)]
-    child.primitive.base_of = None
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = [type(child.primitive)]
-    child.primitive.base_of = [test_aggregation_primitive]
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = None
-    child.primitive.base_of = None
-    child.primitive.base_of_exclude = [test_aggregation_primitive]
-    assert not check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on_exclude = [Count]
-    assert not check_stacking(test_aggregation_primitive(), [child])
-
-    child.primitive.number_output_features = 2
-    test_aggregation_primitive.stack_on_exclude = []
-    test_aggregation_primitive.stack_on = []
-    child.primitive.base_of = []
-    assert not check_stacking(test_aggregation_primitive(), [child])
-
-
-def test_stack_on_self(es, test_aggregation_primitive):
-    # test stacks on self
-    child = Feature(
-        es["log"].ww["value"],
-        parent_dataframe_name="régions",
-        primitive=test_aggregation_primitive,
-    )
-    test_aggregation_primitive.stack_on = []
-    child.primitive.base_of = []
-    test_aggregation_primitive.stack_on_self = False
-    child.primitive.stack_on_self = False
-    assert not check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on_self = True
-    assert check_stacking(test_aggregation_primitive(), [child])
-
-    test_aggregation_primitive.stack_on = None
-    test_aggregation_primitive.stack_on_self = False
-    assert not check_stacking(test_aggregation_primitive(), [child])
 
 
 def test_init_and_name(es):
